Resolve cyclic dependencies.
