+++ src/rich-text-input/RichTextInput.js
@@ -1,13 +1,6 @@
 import {
-  AnnotationComponent,
-  Component,
-  Configurator,
-  EditorSession,
-  AbstractEditor,
-  TextPropertyEditor
+  Component, Configurator, EditorSession, AbstractEditor, TextPropertyEditor
 } from 'substance'
-
-import RichTextInputPackage from './RichTextInputPackage'
 
 export default class RichTextInput extends Component {
   constructor(...args) {
@@ -65,30 +58,5 @@ class RichTextEditor extends AbstractEditor {
       )
     )
     return el
-  }
-}
-
-
-export class StrongComponent extends AnnotationComponent {
-  getTagName() {
-    return 'strong'
-  }
-}
-
-export class EmphasisComponent extends AnnotationComponent {
-  getTagName() {
-    return 'em'
-  }
-}
-
-export class SuperscriptComponent extends AnnotationComponent {
-  getTagName() {
-    return 'sup'
-  }
-}
-
-export class SubscriptComponent extends AnnotationComponent {
-  getTagName() {
-    return 'sub'
   }
 }
